+++ om_account_asset/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     'name': 'Odoo 15 Assets Management',
-    'version': '15.0.5.0.0',
+    'version': '15.0.6.0.0',
     'author': 'Odoo Mates, Odoo SA',
     'depends': ['account'],
     'description': """Manage assets owned by a company or a person.